+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Add disable publication option: javaLib.withoutPublication() (disables all declared publications)
+  Useful for disabling bom publication in the root project (use platform for deps management only) 
+
 ### 2.2.1 (2021-07-13)
 * Fix multi-module projects configuration with allprojects closure 
   (configuration extension always created on plugin activation to avoid mis-references)